requirements.txt
crewai==0.12.1
requests==2.31.0
beautifulsoup4==4.12.3
googlesearch-python==1.2.3
PyYAML==6.0.1
tqdm==4.66.1
langchain==0.1.1
python-dotenv==1.0.0
selenium==4.16.0
chromadb==0.4.22
PyPDF2==3.0.1
SpeechRecognition==3.12.0
# Added missing dependency
